Revert "scale of bitmaps is now float"
This reverts commit cbf3203f761ba94e54dd3c6f0d61abcb90aef854.

+++ IGraphics/IGraphicsStructs.h
@@ -124,7 +124,7 @@ public:
   int N() const { return mN; }
   
   /** @return the scale of the bitmap */
-  float GetScale() const { return mAPIBitmap->GetScale(); }
+  int GetScale() const { return mAPIBitmap->GetScale(); }
 
   /** @return the draw scale of the bitmap */
   float GetDrawScale() const { return mAPIBitmap->GetDrawScale(); }
